Move Kalman file into its own module
